+++ src/Firebase/Messaging/MessageData.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Kreait\Firebase\Messaging;
 
-use Kreait\Firebase\Exception\InvalidArgumentException;
-
 final class MessageData implements \JsonSerializable
 {
     /**
@@ -19,16 +17,13 @@ final class MessageData implements \JsonSerializable
 
     public static function fromArray(array $data): self
     {
+        $messageData = new self();
+
         foreach ($data as $key => $value) {
-            if (!\is_string($key) || !\is_string($value)) {
-                throw new InvalidArgumentException('The keys and values in message data must be all strings.');
-            }
+            $messageData->data[(string) $key] = (string) $value;
         }
 
-        $new = new self();
-        $new->data = $data;
-
-        return $new;
+        return $messageData;
     }
 
     public function jsonSerialize()